refactor: extract storage utilities
